requirements.txt
imaplib2
PySocks
dnspython
pyOpenSSL